requirements.txt
# API
requests==2.31.0

# Telegram
python-telegram-bot==21.1.1

# Timezone
pytz==2024.1

# Templating
Jinja2==3.1.3

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2